+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 # Verb Changelog
+## **unreleased** - unreleased
+- Made Lisp code tags more useful: the HTTP method for a request spec can now be determined by the result of a code tag.
+
 ## **1.0.0** - 2020-01-13
 - Initial release.